generator/requirements.txt
redis==5.0.0
llama-cpp-python==0.2.23
python-dotenv==1.0.0
python-dotenv==1.0.0